refactor: Split components
